+++ src/components/index.ts
@@ -22,6 +22,7 @@ import PMessage from './Message';
 import PToast from './Toast';
 import PList from './List';
 import PLi from './ListItem';
+import PFiguresCard from './FiguresCard';
 import PInfoCard from './InfoCard';
 import PBrandGradientText from './BrandGradientText';
 import PSwitchToggle from './SwitchToggle';
@@ -64,6 +65,7 @@ const components: PhotonComponents = {
   PToast,
   PList,
   PLi,
+  PFiguresCard,
   PInfoCard,
   PBrandGradientText,
   PForm,